Consolidate duplicate subsystems
Merge model-router and smart-router into router.ts; move Docker and
Seatbelt sandboxes behind one sandbox/ interface with mode routing;
unify checkpoints on the git-based mechanism (refs under
refs/calliope/checkpoints, clean no-ops outside git repos); fold
openai-compat-shims into providers/compat.ts. Removes a dead
require()-based file-snapshot call in tools.ts that silently no-opped
under ESM.

ref #181
