+++ src/ui/index.tsx
@@ -29,8 +29,8 @@ import { HUDFrame } from './frame.js';
 import { setEmojiConfig } from '../styles.js';
 import { CircuitBreaker } from '../circuit-breaker.js';
 import { IterationLedger } from '../iteration-ledger.js';
-import { getDefaultSmartRoutingConfig } from '../smart-router.js';
-import type { SmartRoutingConfig } from '../smart-router.js';
+import { getDefaultSmartRoutingConfig } from '../router.js';
+import type { SmartRoutingConfig } from '../router.js';
 import * as sessionTimeout from '../session-timeout.js';
 import * as idleEviction from '../idle-eviction.js';
 import { scuttlebotClient } from '../scuttlebot/index.js';
