+++ tests/checkpoint.test.ts
@@ -1,267 +1,518 @@
 /**
- * Tests for src/checkpoint.ts
+ * Tests for src/checkpoint.ts (unified git-based checkpoints, #181)
  *
- * Covers: createCheckpoint, restoreCheckpoint, listCheckpoints,
- * clearCheckpoints, getCheckpointDir, and edge cases.
- * Uses real temp dirs with os.homedir() mocked to isolate checkpoint storage.
+ * Uses vi.mock('child_process') to avoid real git calls. Covers:
+ *  - shouldCheckpoint / createCheckpoint / revertToLastCheckpoint
+ *  - getCheckpointStats / setEnabled
+ *  - listCheckpoints / restoreFromCheckpoint / clearCheckpoints (ref-based)
+ *  - clean no-ops outside a git work tree
  */
 
-import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
+import { describe, it, expect, vi, beforeEach } from 'vitest';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as os from 'os';
 
-// We need to mock os.homedir() BEFORE importing checkpoint so that
-// getCheckpointDir() uses our temp directory.
-let fakeHome: string;
-
-vi.mock('os', async () => {
-  const actual = await vi.importActual<typeof import('os')>('os');
-  return {
-    ...actual,
-    homedir: () => fakeHome,
-  };
-});
-
-// Import after mock is set up
-import {
-  createCheckpoint,
-  restoreCheckpoint,
-  listCheckpoints,
-  clearCheckpoints,
-  getCheckpointDir,
-} from '../src/checkpoint.js';
-
 // ---------------------------------------------------------------------------
-// Setup / teardown
+// Mock child_process at the module level (ESM-safe approach)
 // ---------------------------------------------------------------------------
 
-let tmpDir: string;
+const mockExecFileSync = vi.fn();
 
-beforeEach(() => {
-  tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'calliope-checkpoint-test-'));
-  fakeHome = tmpDir;
+vi.mock('child_process', () => ({
+  execFileSync: mockExecFileSync,
+}));
+
+// ---------------------------------------------------------------------------
+// Module reset helper — the module carries module-level state.
+// We import after mocking so it picks up our mock.
+// ---------------------------------------------------------------------------
+
+let shouldCheckpoint: typeof import('../src/checkpoint.js').shouldCheckpoint;
+let createCheckpoint: typeof import('../src/checkpoint.js').createCheckpoint;
+let revertToLastCheckpoint: typeof import('../src/checkpoint.js').revertToLastCheckpoint;
+let getCheckpointStats: typeof import('../src/checkpoint.js').getCheckpointStats;
+let setEnabled: typeof import('../src/checkpoint.js').setEnabled;
+let listCheckpoints: typeof import('../src/checkpoint.js').listCheckpoints;
+let restoreFromCheckpoint: typeof import('../src/checkpoint.js').restoreFromCheckpoint;
+let clearCheckpoints: typeof import('../src/checkpoint.js').clearCheckpoints;
+
+beforeEach(async () => {
+  vi.resetModules();
+  mockExecFileSync.mockReset();
+  const mod = await import('../src/checkpoint.js');
+  shouldCheckpoint = mod.shouldCheckpoint;
+  createCheckpoint = mod.createCheckpoint;
+  revertToLastCheckpoint = mod.revertToLastCheckpoint;
+  getCheckpointStats = mod.getCheckpointStats;
+  setEnabled = mod.setEnabled;
+  listCheckpoints = mod.listCheckpoints;
+  restoreFromCheckpoint = mod.restoreFromCheckpoint;
+  clearCheckpoints = mod.clearCheckpoints;
 });
 
-afterEach(() => {
-  fs.rmSync(tmpDir, { recursive: true, force: true });
-});
+// ---------------------------------------------------------------------------
+// shouldCheckpoint
+// ---------------------------------------------------------------------------
 
-// ===========================================================================
-// getCheckpointDir
-// ===========================================================================
-
-describe('getCheckpointDir', () => {
-  it('should return a path under fake home', () => {
-    const dir = getCheckpointDir();
-    expect(dir.startsWith(fakeHome)).toBe(true);
-    expect(dir).toContain('.calliope-cli');
-    expect(dir).toContain('checkpoints');
+describe('shouldCheckpoint', () => {
+  it('should return false when disabled', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    setEnabled(false);
+    expect(shouldCheckpoint('write_file', { path: '/tmp/test.txt' })).toBe(false);
   });
 
-  it('should create the directory if it does not exist', () => {
-    const dir = getCheckpointDir();
-    expect(fs.existsSync(dir)).toBe(true);
+  it('should return false when not in a git repo', () => {
+    mockExecFileSync.mockImplementation(() => { throw new Error('not a git repo'); });
+    expect(shouldCheckpoint('write_file', { path: '/tmp/test.txt' })).toBe(false);
+  });
+
+  it('should return true for write_file in a git repo', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('write_file', { path: '/tmp/test.txt' })).toBe(true);
+  });
+
+  it('should return true for shell command with rm', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'rm somefile.txt' })).toBe(true);
+  });
+
+  it('should return true for shell command with mv', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'mv file.txt file2.txt' })).toBe(true);
+  });
+
+  it('should return true for shell command with cp -f', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'cp -f source.txt dest.txt' })).toBe(true);
+  });
+
+  it('should return true for shell command with git reset', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'git reset --hard HEAD' })).toBe(true);
+  });
+
+  it('should return true for shell command with git clean', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'git clean -fd' })).toBe(true);
+  });
+
+  it('should return true for shell command with sed -i', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'sed -i s/foo/bar/g file.txt' })).toBe(true);
+  });
+
+  it('should return true for shell command with output redirection', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'echo hello > file.txt' })).toBe(true);
+  });
+
+  it('should return true for shell command with dd', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'dd if=/dev/zero of=file' })).toBe(true);
+  });
+
+  it('should return true for shell command with truncate', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'truncate -s 0 file.txt' })).toBe(true);
+  });
+
+  it('should return false for safe shell command', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', { command: 'echo hello' })).toBe(false);
+  });
+
+  it('should return false for unknown tool', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('read_file', { path: '/tmp/test.txt' })).toBe(false);
+  });
+
+  it('should handle missing command arg gracefully', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('shell', {})).toBe(false);
   });
 });
 
-// ===========================================================================
+// ---------------------------------------------------------------------------
 // createCheckpoint
-// ===========================================================================
+// ---------------------------------------------------------------------------
+
+/** Full-success mock producing the given short hash. */
+function mockSuccessfulCheckpoint(hash: string, capture?: string[][]) {
+  mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+    if (args[0] === 'rev-parse' && args[1] === '--is-inside-work-tree') return 'true\n';
+    if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+    if (args[0] === 'status') return 'M file.txt\n';
+    if (args[0] === 'add') return '';
+    if (args[0] === 'diff') return 'some staged diff\n';
+    if (args[0] === 'commit') return '';
+    if (args[0] === 'rev-parse' && args[1] === '--short') return `${hash}\n`;
+    if (args[0] === 'update-ref') { capture?.push(args); return ''; }
+    if (args[0] === 'reset') return '';
+    return '';
+  });
+}
 
 describe('createCheckpoint', () => {
-  it('should save provided content and return a filename', () => {
-    const filePath = path.join(tmpDir, 'file.txt');
-    const filename = createCheckpoint(filePath, 'original content');
-
-    expect(filename).toBeDefined();
-    expect(typeof filename).toBe('string');
-    expect(filename!.startsWith('checkpoint-')).toBe(true);
-    expect(filename!.endsWith('.json')).toBe(true);
-
-    // Verify the checkpoint file was written
-    const cpPath = path.join(getCheckpointDir(), filename!);
-    expect(fs.existsSync(cpPath)).toBe(true);
-
-    const cp = JSON.parse(fs.readFileSync(cpPath, 'utf-8'));
-    expect(cp.content).toBe('original content');
-    expect(cp.filePath).toBe(path.resolve(filePath));
-    expect(cp.timestamp).toBeDefined();
+  it('should return null when not in a git repo', () => {
+    mockExecFileSync.mockImplementation(() => { throw new Error('not a git repo'); });
+    expect(createCheckpoint('write_file', { path: 'test.txt' })).toBeNull();
   });
 
-  it('should read content from disk when not provided', () => {
-    const filePath = path.join(tmpDir, 'ondisk.txt');
-    fs.writeFileSync(filePath, 'disk content');
-
-    const filename = createCheckpoint(filePath);
-    expect(filename).toBeDefined();
-
-    const cpPath = path.join(getCheckpointDir(), filename!);
-    const cp = JSON.parse(fs.readFileSync(cpPath, 'utf-8'));
-    expect(cp.content).toBe('disk content');
+  it('should return null when git status is empty (nothing to checkpoint)', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse') return 'true\n';
+      if (args[0] === 'status') return ''; // empty = nothing to commit
+      return '';
+    });
+    expect(createCheckpoint('write_file', { path: 'test.txt' })).toBeNull();
   });
 
-  it('should return undefined when file does not exist and no content given', () => {
-    const result = createCheckpoint(path.join(tmpDir, 'ghost.txt'));
-    expect(result).toBeUndefined();
+  it('should return null when nothing is staged after git add', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse') return 'true\n';
+      if (args[0] === 'status') return 'M modified.txt\n';
+      if (args[0] === 'add') return '';
+      if (args[0] === 'diff') return ''; // nothing staged
+      return '';
+    });
+    expect(createCheckpoint('write_file', { path: 'test.txt' })).toBeNull();
   });
 
-  it('should store optional sessionId', () => {
-    const filePath = path.join(tmpDir, 'session.txt');
-    const filename = createCheckpoint(filePath, 'data', 'session-123');
-
-    const cpPath = path.join(getCheckpointDir(), filename!);
-    const cp = JSON.parse(fs.readFileSync(cpPath, 'utf-8'));
-    expect(cp.sessionId).toBe('session-123');
+  it('should create checkpoint for write_file and return hash', () => {
+    mockSuccessfulCheckpoint('abc1234');
+    expect(createCheckpoint('write_file', { path: 'src/main.ts' })).toBe('abc1234');
   });
 
-  it('should not include sessionId when not provided', () => {
-    const filePath = path.join(tmpDir, 'nosession.txt');
-    const filename = createCheckpoint(filePath, 'data');
-
-    const cpPath = path.join(getCheckpointDir(), filename!);
-    const cp = JSON.parse(fs.readFileSync(cpPath, 'utf-8'));
-    expect(cp.sessionId).toBeUndefined();
-  });
-});
-
-// ===========================================================================
-// restoreCheckpoint
-// ===========================================================================
-
-describe('restoreCheckpoint', () => {
-  it('should restore the most recent checkpoint', async () => {
-    const filePath = path.join(tmpDir, 'restore-me.txt');
-
-    // Create two checkpoints with slight delay
-    createCheckpoint(filePath, 'version 1');
-    // Ensure different timestamps
-    await new Promise(r => setTimeout(r, 10));
-    createCheckpoint(filePath, 'version 2');
-
-    // Write something else to the file
-    fs.writeFileSync(filePath, 'current content');
-
-    // Restore index 0 = most recent = version 2
-    const restored = restoreCheckpoint(filePath, 0);
-    expect(restored).toBe('version 2');
-    expect(fs.readFileSync(filePath, 'utf-8')).toBe('version 2');
+  it('should create checkpoint for shell command and return hash', () => {
+    mockSuccessfulCheckpoint('def5678');
+    expect(createCheckpoint('shell', { command: 'rm -rf /tmp/test' })).toBe('def5678');
   });
 
-  it('should restore an older checkpoint by index', async () => {
-    const filePath = path.join(tmpDir, 'restore-old.txt');
-
-    createCheckpoint(filePath, 'v1');
-    await new Promise(r => setTimeout(r, 10));
-    createCheckpoint(filePath, 'v2');
-
-    const restored = restoreCheckpoint(filePath, 1);
-    expect(restored).toBe('v1');
+  it('should create checkpoint for unknown tool using tool name as summary', () => {
+    mockSuccessfulCheckpoint('bbb9999');
+    expect(createCheckpoint('some_other_tool', {})).toBe('bbb9999');
   });
 
-  it('should return undefined when no checkpoints exist', () => {
-    const result = restoreCheckpoint(path.join(tmpDir, 'nocheckpoint.txt'));
-    expect(result).toBeUndefined();
+  it('should return null when git commit throws', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--is-inside-work-tree') return 'true\n';
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'status') return 'M file.txt\n';
+      if (args[0] === 'add') return '';
+      if (args[0] === 'diff') return 'some diff\n';
+      if (args[0] === 'commit') throw new Error('commit failed');
+      return '';
+    });
+    expect(createCheckpoint('write_file', { path: 'test.txt' })).toBeNull();
   });
 
-  it('should return undefined for out-of-range index', () => {
-    const filePath = path.join(tmpDir, 'oor.txt');
-    createCheckpoint(filePath, 'data');
-
-    expect(restoreCheckpoint(filePath, 5)).toBeUndefined();
-    expect(restoreCheckpoint(filePath, -1)).toBeUndefined();
+  it('should increment checkpoint count on success', () => {
+    mockSuccessfulCheckpoint('aaa1111');
+    createCheckpoint('write_file', { path: 'test.txt' });
+    const stats = getCheckpointStats();
+    expect(stats.count).toBe(1);
+    expect(stats.lastHash).toBe('aaa1111');
   });
 
-  it('should create parent directories if needed', () => {
-    const filePath = path.join(tmpDir, 'deep', 'nested', 'file.txt');
-    createCheckpoint(filePath, 'deep content');
+  it('records a checkpoint ref pointing at the commit', () => {
+    const updateRefCalls: string[][] = [];
+    mockSuccessfulCheckpoint('abc1234', updateRefCalls);
+    expect(createCheckpoint('write_file', { path: 'a.txt' })).toBe('abc1234');
+    expect(updateRefCalls.length).toBe(1);
+    expect(updateRefCalls[0][0]).toBe('update-ref');
+    expect(updateRefCalls[0][1]).toMatch(/^refs\/calliope\/checkpoints\//);
+    expect(updateRefCalls[0][2]).toBe('abc1234');
+  });
 
-    const restored = restoreCheckpoint(filePath);
-    expect(restored).toBe('deep content');
-    expect(fs.existsSync(filePath)).toBe(true);
+  it('still returns the hash when ref tracking fails (best-effort)', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--is-inside-work-tree') return 'true\n';
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'status') return 'M file.txt\n';
+      if (args[0] === 'add') return '';
+      if (args[0] === 'diff') return 'staged\n';
+      if (args[0] === 'commit') return '';
+      if (args[0] === 'rev-parse' && args[1] === '--short') return 'abc1234\n';
+      if (args[0] === 'update-ref') throw new Error('ref failed');
+      return '';
+    });
+    expect(createCheckpoint('write_file', { path: 'a.txt' })).toBe('abc1234');
   });
 });
 
-// ===========================================================================
+// ---------------------------------------------------------------------------
+// revertToLastCheckpoint
+// ---------------------------------------------------------------------------
+
+describe('revertToLastCheckpoint', () => {
+  it('should return false when no checkpoint has been created', () => {
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(revertToLastCheckpoint()).toBe(false);
+  });
+
+  it('should return true after successfully reverting', () => {
+    mockSuccessfulCheckpoint('abc1234');
+    createCheckpoint('write_file', { path: 'test.txt' });
+    expect(revertToLastCheckpoint()).toBe(true);
+  });
+
+  it('should return false when git reset throws', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--is-inside-work-tree') return 'true\n';
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'status') return 'M file.txt\n';
+      if (args[0] === 'add') return '';
+      if (args[0] === 'diff') return 'some diff\n';
+      if (args[0] === 'commit') return '';
+      if (args[0] === 'rev-parse' && args[1] === '--short') return 'abc1234\n';
+      if (args[0] === 'reset') throw new Error('reset failed');
+      return '';
+    });
+    createCheckpoint('write_file', { path: 'test.txt' });
+    expect(revertToLastCheckpoint()).toBe(false);
+  });
+
+  it('should reset against the captured repo root, not live cwd', () => {
+    const resetCalls: string[][] = [];
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[], opts: { cwd: string }) => {
+      if (args[0] === 'rev-parse' && args[1] === '--is-inside-work-tree') return 'true\n';
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo-a\n';
+      if (args[0] === 'status') return 'M file.txt\n';
+      if (args[0] === 'add') return '';
+      if (args[0] === 'diff') return 'some diff\n';
+      if (args[0] === 'commit') return '';
+      if (args[0] === 'rev-parse' && args[1] === '--short') return 'abc1234\n';
+      if (args[0] === 'update-ref') return '';
+      if (args[0] === 'reset') { resetCalls.push([opts.cwd, ...args]); return ''; }
+      return '';
+    });
+    createCheckpoint('write_file', { path: 'test.txt' });
+    expect(revertToLastCheckpoint()).toBe(true);
+    expect(resetCalls).toEqual([['/repo-a', 'reset', '--hard', 'abc1234']]);
+  });
+
+  it('should be a no-op when the stored repo root no longer matches', () => {
+    let toplevel = '/repo-a\n';
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--is-inside-work-tree') return 'true\n';
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return toplevel;
+      if (args[0] === 'status') return 'M file.txt\n';
+      if (args[0] === 'add') return '';
+      if (args[0] === 'diff') return 'some diff\n';
+      if (args[0] === 'commit') return '';
+      if (args[0] === 'rev-parse' && args[1] === '--short') return 'abc1234\n';
+      if (args[0] === 'update-ref') return '';
+      if (args[0] === 'reset') throw new Error('should not reset on mismatch');
+      return '';
+    });
+    createCheckpoint('write_file', { path: 'test.txt' });
+    toplevel = '/repo-b\n'; // cwd moved to a different repo before the revert
+    expect(revertToLastCheckpoint()).toBe(false);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// getCheckpointStats / setEnabled
+// ---------------------------------------------------------------------------
+
+describe('getCheckpointStats', () => {
+  it('should return initial stats', () => {
+    const stats = getCheckpointStats();
+    expect(stats.enabled).toBe(true);
+    expect(stats.count).toBe(0);
+    expect(stats.lastHash).toBeNull();
+  });
+
+  it('should reflect disabled state after setEnabled(false)', () => {
+    setEnabled(false);
+    expect(getCheckpointStats().enabled).toBe(false);
+  });
+
+  it('should reflect enabled state after setEnabled(true)', () => {
+    setEnabled(false);
+    setEnabled(true);
+    expect(getCheckpointStats().enabled).toBe(true);
+  });
+});
+
+// ---------------------------------------------------------------------------
+// isGitRepo re-evaluation (must not memoize across cwd changes)
+// ---------------------------------------------------------------------------
+
+describe('isGitRepo re-evaluation', () => {
+  it('should re-evaluate git-repo status on each call (no stale memoization)', () => {
+    mockExecFileSync.mockImplementationOnce(() => { throw new Error('not a git repo'); });
+    expect(shouldCheckpoint('write_file', { path: 'test.txt' })).toBe(false);
+
+    mockExecFileSync.mockReturnValue('true\n');
+    expect(shouldCheckpoint('write_file', { path: 'test.txt' })).toBe(true);
+
+    mockExecFileSync.mockImplementation(() => { throw new Error('not a git repo'); });
+    expect(shouldCheckpoint('write_file', { path: 'test.txt' })).toBe(false);
+  });
+});
+
+// ---------------------------------------------------------------------------
 // listCheckpoints
-// ===========================================================================
+// ---------------------------------------------------------------------------
+
+const REF_A = 'refs/calliope/checkpoints/2026-07-04T17-00-02-000Z-0002';
+const REF_B = 'refs/calliope/checkpoints/2026-07-04T17-00-01-000Z-0001';
+// for-each-ref --sort=-refname emits newest (REF_A) first.
+const FOR_EACH_REF_OUT =
+  `hashaaa\t${REF_A}\t2026-07-04T17:00:02-07:00\t[checkpoint] before write_file: a.txt\n` +
+  `hashbbb\t${REF_B}\t2026-07-04T17:00:01-07:00\t[checkpoint] before write_file: b.txt\n`;
 
 describe('listCheckpoints', () => {
-  it('should return empty list when no checkpoints exist', () => {
+  it('returns empty list outside a git work tree', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') throw new Error('not a repo');
+      return '';
+    });
     expect(listCheckpoints()).toEqual([]);
   });
 
-  it('should list all checkpoints', async () => {
-    const f1 = path.join(tmpDir, 'a.txt');
-    const f2 = path.join(tmpDir, 'b.txt');
-
-    createCheckpoint(f1, 'a');
-    await new Promise(r => setTimeout(r, 10));
-    createCheckpoint(f2, 'b');
-
+  it('lists all checkpoints newest-first with parsed fields', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'for-each-ref') return FOR_EACH_REF_OUT;
+      return '';
+    });
     const all = listCheckpoints();
     expect(all.length).toBe(2);
+    expect(all[0]).toEqual({
+      ref: REF_A,
+      hash: 'hashaaa',
+      timestamp: '2026-07-04T17:00:02-07:00',
+      subject: '[checkpoint] before write_file: a.txt',
+    });
+    expect(all[1].hash).toBe('hashbbb');
   });
 
-  it('should filter by file path', async () => {
-    const f1 = path.join(tmpDir, 'x.txt');
-    const f2 = path.join(tmpDir, 'y.txt');
-
-    createCheckpoint(f1, 'x');
-    await new Promise(r => setTimeout(r, 10));
-    createCheckpoint(f2, 'y');
-
-    const filtered = listCheckpoints(f1);
+  it('filters to checkpoints whose tree contains the file', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'for-each-ref') return FOR_EACH_REF_OUT;
+      if (args[0] === 'cat-file') {
+        // Only the newest checkpoint (hashaaa) contains the file.
+        if (args[2] === 'hashaaa:wanted.txt') return '';
+        throw new Error('not in tree');
+      }
+      return '';
+    });
+    const filtered = listCheckpoints('/repo/wanted.txt');
     expect(filtered.length).toBe(1);
-    expect(filtered[0].filePath).toBe(path.resolve(f1));
+    expect(filtered[0].hash).toBe('hashaaa');
   });
 
-  it('should sort newest first', async () => {
-    const filePath = path.join(tmpDir, 'sorted.txt');
-
-    createCheckpoint(filePath, 'older');
-    await new Promise(r => setTimeout(r, 15));
-    createCheckpoint(filePath, 'newer');
-
-    const list = listCheckpoints(filePath);
-    expect(list.length).toBe(2);
-    // First entry should have a later timestamp
-    expect(list[0].timestamp >= list[1].timestamp).toBe(true);
-  });
-
-  it('should include size in summaries', () => {
-    const filePath = path.join(tmpDir, 'sized.txt');
-    createCheckpoint(filePath, 'hello world');
-
-    const list = listCheckpoints(filePath);
-    expect(list[0].size).toBe('hello world'.length);
+  it('returns empty when for-each-ref produces no output', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'for-each-ref') return '';
+      return '';
+    });
+    expect(listCheckpoints()).toEqual([]);
   });
 });
 
-// ===========================================================================
-// clearCheckpoints
-// ===========================================================================
+// ---------------------------------------------------------------------------
+// restoreFromCheckpoint
+// ---------------------------------------------------------------------------
 
-describe('clearCheckpoints', () => {
-  it('should remove all checkpoints when no days specified', () => {
-    createCheckpoint(path.join(tmpDir, 'a.txt'), 'a');
-    createCheckpoint(path.join(tmpDir, 'b.txt'), 'b');
+describe('restoreFromCheckpoint', () => {
+  let tmpDir: string;
 
-    const removed = clearCheckpoints();
-    expect(removed).toBe(2);
-    expect(listCheckpoints().length).toBe(0);
+  beforeEach(() => {
+    tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), 'calliope-cp-restore-'));
   });
 
-  it('should return 0 when no checkpoints exist', () => {
+  it('restores a file from the most recent checkpoint and writes it to disk', () => {
+    const target = path.join(tmpDir, 'file.txt');
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return `${tmpDir}\n`;
+      if (args[0] === 'for-each-ref') {
+        return `hashaaa\t${REF_A}\t2026-07-04T17:00:02-07:00\t[checkpoint] before write_file: file.txt\n`;
+      }
+      if (args[0] === 'cat-file') return ''; // file present in tree
+      if (args[0] === 'show') return 'restored content';
+      return '';
+    });
+
+    const restored = restoreFromCheckpoint(target, 0);
+    expect(restored).toBe('restored content');
+    expect(fs.readFileSync(target, 'utf-8')).toBe('restored content');
+  });
+
+  it('returns undefined for an out-of-range index', () => {
+    const target = path.join(tmpDir, 'file.txt');
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return `${tmpDir}\n`;
+      if (args[0] === 'for-each-ref') {
+        return `hashaaa\t${REF_A}\t2026-07-04T17:00:02-07:00\t[checkpoint] before write_file: file.txt\n`;
+      }
+      if (args[0] === 'cat-file') return '';
+      return '';
+    });
+    expect(restoreFromCheckpoint(target, 5)).toBeUndefined();
+    expect(restoreFromCheckpoint(target, -1)).toBeUndefined();
+  });
+
+  it('returns undefined outside a git work tree', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') throw new Error('not a repo');
+      return '';
+    });
+    expect(restoreFromCheckpoint(path.join(tmpDir, 'file.txt'))).toBeUndefined();
+  });
+});
+
+// ---------------------------------------------------------------------------
+// clearCheckpoints
+// ---------------------------------------------------------------------------
+
+describe('clearCheckpoints', () => {
+  it('returns 0 outside a git work tree', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') throw new Error('not a repo');
+      return '';
+    });
     expect(clearCheckpoints()).toBe(0);
   });
 
-  it('should keep recent checkpoints when olderThanDays is specified', () => {
-    // Create a checkpoint (just now, so within any reasonable days threshold)
-    createCheckpoint(path.join(tmpDir, 'recent.txt'), 'recent');
+  it('deletes all checkpoint refs when no age is given', () => {
+    const deleted: string[] = [];
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'for-each-ref') return `${REF_A}\t1751600002\n${REF_B}\t1751600001\n`;
+      if (args[0] === 'update-ref' && args[1] === '-d') { deleted.push(args[2]); return ''; }
+      return '';
+    });
+    expect(clearCheckpoints()).toBe(2);
+    expect(deleted).toEqual([REF_A, REF_B]);
+  });
 
-    const removed = clearCheckpoints(1); // Older than 1 day
-    expect(removed).toBe(0);
-    expect(listCheckpoints().length).toBe(1);
+  it('keeps recent checkpoints when olderThanDays is specified', () => {
+    const nowUnix = Math.floor(Date.now() / 1000);
+    const oldUnix = nowUnix - 10 * 24 * 60 * 60; // 10 days ago
+    const deleted: string[] = [];
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'for-each-ref') return `${REF_A}\t${nowUnix}\n${REF_B}\t${oldUnix}\n`;
+      if (args[0] === 'update-ref' && args[1] === '-d') { deleted.push(args[2]); return ''; }
+      return '';
+    });
+    // Only the 10-day-old checkpoint is older than 1 day.
+    expect(clearCheckpoints(1)).toBe(1);
+    expect(deleted).toEqual([REF_B]);
+  });
+
+  it('returns 0 when there are no checkpoint refs', () => {
+    mockExecFileSync.mockImplementation((_cmd: unknown, args: string[]) => {
+      if (args[0] === 'rev-parse' && args[1] === '--show-toplevel') return '/repo\n';
+      if (args[0] === 'for-each-ref') return '';
+      return '';
+    });
+    expect(clearCheckpoints()).toBe(0);
   });
 });
